Upgrade to unicorn hat hd
