+++ app/programs/hd/__init__.py
@@ -1,0 +1,1 @@
+from . import candle, demo, forest_fire, game_of_life, rainbow, stars
